Switch to using PIO for the display to reduce flicker
This uses code from the hub75 example in the pico-examples repo.

This involved switching the framebuffer around so it is y, x now. The
row and row+16 are copied in a criticial section, insetead of using a
semaphore, which seems to reduce flicker further.

I may look at DMA at some point, but for now the screen is good enough
and it's time to get on with displaying stuff.