Added bigger weather icons and a Go program for generating them
The Go program is a bit rough and ready, but it works. Uses
ImageMagick for the cropping and scaling of a royalty-free
image containing 16 weather icons, which roughly map to Home
Assistant's "weather conditions".

Added a basic image gamma control, useable when drawing images.
This is used to reduce the intensity of the large weather icon
on the current weather stats page.

Implemented a basic display for the current weather conditions, and
then started playing with a scrolling message for additional weather
info, but it might actually be useable for displaying all kinds of
random facts about what HA knows about the home, which could be shown,
periodically, on all screens.