Added a brightness topic
0-255. Propogates via a message to the animator task. Actual
implemenntation is via the framebuffer::atomic_back_to_fore_copy()
method. In the process of doing this the fb backend datastructure
is now a union of two arrays: one of rgb_t and one of uint32_t,
in order to make use of the underlying data, as needed by the PIO
output, simpler.

Added some underscores to various framebuffer draw methods while
I'm here.