Made "transparent" iamges optional
On a notification, the background is drawn in a light colour instead of
black, initially, to flash the screen. This means images, like clouds
etc, needed to be drawn with black being "transparent" ie, not set into
the image for their black background. The problem with that is the
porch image needs to be not transparent for the PORCH writing to be
leggible. So make transparency optional.