Added a parser for the local bus timetables
Totally useless to anyone else, but useful to me. It pokes the realtime
bus information into MQTT, which will be picked out by the display.

To save load on the target server the request is only made if the
display is on.

This service will eventually be wrapped in some kind of container and
will run on the same host as the Home Assistant and MQTT broker.