Add bus times, switch clock to generic DS1307 alike, redo statemachine
Added initial bus times for the bluestar busses using mostly
code that should be factored into common code.

Changed the horrible render_page() state machine to use the renderers
themselves to determine if the next page should be shown instead of
scattering this code everywhere, which made it much harder to add
new pages. The cost is one render cycle when a page is switched to
which does not have anything to display.