Added grayscale and per R G B brightness
For Halloween, make it so the screen has individual RGB brightness levels.
Also made it so the pixel colour can be averaged out via a "greyscale"
mode.

Added temp from RTC IC tot he weather screen, though I thought this was
in an earlier commit.

Tweaked logging messages as I now have this thing on the mantlepiece
"in production" with a log capture board attached.