Added configuration fields
Made it possible to configure the number of refreshes each "page" should
appear for, as well as the speed (actually the delay count) of ssome
scrolling messages. You can hide a page by settings its duration field
to 0. These values are all under matrix_display/confiugraiton/X.

Also updated the bus parser script to cater for the removal of the overall
brightness field, which I forgot to include in the previous commit.