Added winter-themed snowflakes
These are controlled via another configuration variable, snowflake_count.
0 to 255 snowflakes. :)

Simple animation, with wind. The snowflakes use a new field in the unused
byte of the 32 bit RGB value, which causes the brightness to be ignored
when copying a pixel with that flag set.

Also fixed some nonsense in the change to the RTC time/date formatting.
The rendering of the bus info in red is still not tested though. It might
infact be useful to apply that logic to the weather display too.