Added MQTT autodiscover support
This is pretty neat. The end user no longer has to manually add entities
to their configuration.yaml file! Instead a new topic is exposed:

matrix_display/autodiscover

And if ON is written to this then the device will reply with the needed
pubs (/homeassistant/$class/matrix_display/config) which will create
the entities and the device. Currently the following is exposed:

light.matrix_display_panel : a dimmable light representing the panel
switch.matrix_display_grayscale : a switch for gray or not
number.matrix_display_snowflakes : the count of snowflakes (0-255)
sensor.matrix_display_temperature : the current temp

The snowflake count is actaully under the configuration topic tree
(matrix_display/configuration/snowflake_count) but only the snowflake
count is currently exposed via this method. The old config values are
still there, but most be poked in via MQTT.

Note that this currently needs a bodge into the MQTT LwIP code:

Without this the pubs will be too big. I haven't yet figured out how
to set this value from this code tree.