Made the choice of FPGA or direct HUB75 connection a compile time choice
Since the latest board supports both an iCE40UP5 FPGA and a direct
connection between the Pico and the HUB75 header, lets support both
in a single board.

Therefore someone can build a board and populate it with an FPGA or
not, and use the same repo clone.

Compile time options should probably all come from one place, the
commandline, since this is where the Wi-Fi details will come from.

Nearly ready for our first "release".