Added support for BME680 sensor
This is a neat sensor that supoprts temperature, pressure, humidty
and an air quality index value. The later requires a closed source
library annoyingly. For now, just expose the other 3 values.

Support for this device is optional via -DBME680_PRESENT=1 on the
cmake commandline. Without it, temperature data will continue to
come from the DS3231 RTC IC. Note that the I2C slave address is
assumed to be 0x76, which requires the SDO pin to be tied low.

Renamed references to RTC to DS3231, which is hopefully clearer:
the format of the time/date packet is afterall specific to this
IC. I may yet abstract away the real time clock mechanism, in a
simar way to how "climate" is done here. It's not ever C++'y but
as this is build-time it sort of makes sense to approach it this
way.

TODO: elimate some repetition from the publishing of the sensor
JSON blocks, and other autodiscovery packets.