Extend the scroller concept to an array of strings passed via MQTT
The scrollers were a bit unused and ineffective: previously they were
rendered into text in the firmware, instead pass them in via MQTT with
formatting done in Home Assistant or whatever. Currently the frimware
holds 4 messages, which are displayed in a cycle. The current weather
message is now formed in my HA automation. The second message is used
for displaying stock prices, which I'm getting via Yahoo using a 3rd
party component. The strings are passed in via the MQTT topic
matrix_display/scroller which wants an array of up to four strings.

In order to show arrows for a price going up or down, the
framebuffer::print_char() call has been extended: two new chars
have been added at 0x80 and 0x81 for arrows, and the colour for that
character only is forced to either red or green. Further, to propogate
those bytes into the string a replacement function has been added to
fonts.c to replace -UP- and -DOWN with those bytes. (I really did not
fancy looking at UTF-8 at all, and it seems JSON does not include
embedded arbitary character values in its strings like C does, though
this solution is pretty horrible it is the least bad.)

Also made the colons in the time display blink, using a new config
paramater clock_colon_flash, which is the count of frames to show
the clock before toggling the colons. 0 to make it revert to the old
way, which is the default. 20 seems to work nicely.

Added the chance of rain to the current weather conditions screen,
which fills it out nicely and stops it being quite so barron.

Finally, the usual small cleanups. There's naturally a large scope
for refactoring. But... that's boring. :)

TODO: update README.md.