Added a watchdog
This will reset the MCU after 2 seconds of the last poke, which
happens during the RTC tick handler in the animation task.

Shouldn't be necessary. :( But I've been working on trying to resolve
a crash in the MQTT task for ages without success.

Not a great solution at all, but I'm telling myself that nearly all
non trivial MCU software has them to fix "issues".