Made screens transition via fade outs and fande ins
Added a bunch more state variables to make it so the individual pages
will not just flip in and out, but smoothly transition. Works quite
well! Added a rgb_faded() function to scale a colour by a uint8_t.
The scaling is used everywhere except in a few places like the scrollers
and notifications which we don't want to fade out as the main screen
switches page.

Scrolling pages do not really transition out, that is they do not
(currently) start dimming before the last bit of text has left the
screen. A nicer example is when the current weather transitions to the
weather forecast.

Obviously the framebuffer brightness, as manipulated by the
brightness MQTT topic is another layer of intensity.

Might make the fade interface (FADE_DURATION currently at 32) into
a configuration parameter, dunno yet.