Added RTTTL parsing and a playback MQTT topic
Uses a class around a couple of functions that parse a RTTTL "rintone"
tune. The old beeps, used by notification and porch activity have
been renamed "simple". They are not available to MQTT anymore, as they
are redundant really. It is also possible to make a notification
(critical or not) have an accompaning tune. If supplied it will be
played instead of the standard beeps. This is untested.