Very early days for an ESP32 port
This is fun. I have an "Matrix Portal S3" board and I Thought it would
be interesting to try to get this project ported to it.

An envvar, MATRIX_DISPLAY_SDK, controls the build: PICO for the old Pico W
based build and ESP32 for... ESP32. This required moving the source under
a dir off the project root. ESP-IDF generally use main/ for this but with
some work you can all it something else, so I went for src/ to match most
normal repos where docs etc sit about the code.

Currently you can still build the old Pico W codebase, but ESP32 is
nowhere near finished. The next thing to work on will be the MQTT task,
which will include networking.

For the display itself I have several options, but to keep it simple for
now I will bitbang the HUB75 pins. With luck this will be sufficent and
I wont have to look at other, more exotic, choices....