Further work on ESP32 port
The port mostly runs! No display yet, but WIFI connects and MQTT too.

There may yet be a centrallised "HAL" instead of spreading #ifs around the
place, not sure yet.

Pico W build is sort of broken. It builds fine, but no USB TTY is created
and nothing is on the screen. That's probably the next thing to get working
again, or I might focus on bitbaning an image out the ESP32.

fonts/ and images/ have been simplfied and are now just more sources buid
out of the first CMAkeFile. Was easier doing that then wrestling with local
include paths for sub directories.

Otherwise, making slow but steady progress.