More work on the ESP32 port
Mostly works! Time is sync'd via SNTP; an RTC IC is no longer required.
This will be the case for all target boards.

The actual display is bit-banged. This required some extra code to disable
the watchdog on the second core, since a busy looping core will trip the
dog by default. This is daft, but there. Also LWIP only operates on the
first core. This required some menuconfig change. No flickering has been
observed since making that last change, which is great.

MQTT subs now listen only on a couple of wildcards. Also, since this
house might end up with multiple matrix-panels in it, the topics are now
both house-wide and instance wide. House wide for things like the weather
and instance wide for things like notifications.

All needs some work to tidy up the code, but before doing that I want to
get the Pico build working again, with SNTP etc.