Make SNTP work on Pico too
Was fairly easy in the end. The bulk of the calls are the same, only
the callback is different. A macro sets a function name to call, which
is passed in via CMake. Also in the Pico the FreeRTOS clock is not
automatically updated and this must be done with some code.

Wifi connection on the Pico is still awful. Needs some research to find
the recommended approach.

The last bit of functionality is to get external sensors working, which
will require some kind of abstraction for I2C. Also needed, beside a
massive amount of cleaning up, is a macro for the topics so we can do
things like matrix_display_livingroom/notification etc.