Add an I2C layer, improve Pico WiFI startup and autodetect etc
Added an I2C portability layer so sensors don't have to care if a Pico
or ESP32 is in use. Very simple. ESP32 version not written yet.

Made WiFi startup a bit better by using the async connection function and
looping until IP is available.

Reworked the autodetect (and other publish) calls by moving the
autodetect JSON transmittion to the subscribe callback. Found that
the LwIP MQTT ringbuffer wasn't big enough and mqtt_opts.h infact wasn't
being used, so moved the setting for the ringbuffer size to a define
passed in via CMake.

Lastly because I might have two panels in my HA I made it so the
"pretty name" of the device and the topics used can be preturbed
at build time. eg. matrix_ppanel_lounge/whatever. For now things like
the weather topic is fixed at matrix_panel/weather because otherwise
the weather will have to pushed out multiple times for each panel,
which is probably not desired since the weather for each panel is
probably the same. This is the DEVICE_PRETTY_NAME and DEVICE_POSTFIX
variables. The first defaults to "Matrix Display" and the second is
empty by default, making all topics just matrix_display/something.