Added an I2C layer for ESP-IDF; working with the DS3231 so far
This uses the older i2c.h API, since it much more closely matches the
Pico SDK. The new ESP-IDF I2C API has the concept of a device, which
needs to be created for each peripheral IC that will be contacted,
which seems like a massive overkill for what we need. The I2C "HAL"
has been modified to reflect what ESP-IDF does with write-then-read
transactions. Instead of having a handy "don't send stop" parameter
on the read and write calls it always sends a stop at the end. If you
want to a write followed by a read (eg. you are interfacing with some
kind of memory-like peripheral like an RTC) you must use a 3rd type
of function: write then read. So that is the pattern that is copied
by the "hal" here.

I still need to propogate in the GPIO numbers from defines. Currently
just using two random pins on the Adafruit Matrix Portal S3 board.

Testing MQTT pubs on the ESP32S3 and I have problems with pub'ing the
temp data. Random but very repeatable panics. For now that's been turned
off on the ESP32 build.

Experimenting with LOCK_TCP_CORE() as a ESP32 equivilant to
cyw43_arch_lwip_begin. Not really sure if its needed but it was an
attempt to get publishing working without crashes. Still not working so
will prob take it out.

I have two further sensors to add: one is a boring I2C temp sensor, the
other is an ambient light sensor board. Should be fun.