Make publish be a response to a subscribed topic
Not the best solution for the crashes I'm getting on the ESP32 when doing
mqtt_publish() from outside a callback, but this will do for now. HA
(or whatever) needs to be setup to send a message to
matrix_display_whatever/publish_trigger and the device will respond by
publishing the last value from its sensors.