Reworked sensors to use an abstract "Sensor" class
This is much nicer.  Each sensor type (TemperatureSensor etc) is another
abstract class, which has a pure virtual method to expose the read in value.
The steps for collecting the values with a given sensor IC are:

1. configure() - called when the sensor tasks starts
2. request_run() - called on the penultimate second of each minute
3. receive_data() - called once for the whole sensor at the end of each
minute
4. get_temperature() - called to get the actual value at the end of each
minute, after the above

Only the last one is mandatory.

Some trickery with a method to return a pointer to the object in the
needed class (TemperatureSensor etc) is used since multiple inheritence
isn't really possible where virtual abstract classes are required in non
RTTI setups like what we have here with embedded MCUs. The same method is
used to determine if a sensor has a particular type of data to expose.

Thanks DC!

In a followup commit this mechanism will be extended so multiple
temperature (say) sensors can be combined to expose an average reading.

This commit also adds support the BH1750 light level sensor, though the
interactions with it is basic at the moment.

Lastly all references to "climate" have been removed. It's just sensors
now.