Added a name field to each sensor and added some error checking to I2C
Errors will be reported, including the sensor name. We could, I guess,
remove a failed sensor from the list of sensors but since the problem
might be transitory it's probably not the right move.

Mext thing will be back to boring code cleanups, including a grand
rename of all classes and typedefs. :(