Added board definitions
To make building this for random boards as easy as possible, the per-
board settings have been moved to board definition files. This includes
the pinning information and how the HUB75 should be controlled, in the
case of Pico based boards anyway. And of course the SDK to use to build
the code for that board is included as a board attribute.

Wifi etc details have been moved to this file which sets the board to
build for, and it's now called local.cmake.

The currently defined boards are:

* My own custom board featuring a now mostly redundant ICE40UP5 FPGA
  * This is the only case where the SPI_TO_FPGA mode can be used, so
    currently it is not possible to build a board with an ESP32 and use
    this display method.
* The Pimoroni Interstate 75W, which is an RP23050 based board
  * Screen driven by PIO mode.
* The Adafruit MatrixPortal S3, which is an ESP32 S3 based board
  * Screen bitbanged.

In all cases the I2C bus is defined and sensors can be attached by
setting them in the sensors.cmake file.