Add a starting up progress message system, and fix the light MQTT topics
To show progress on startup send messages from the MQTT task to the
animation task showing things like "WiFi connecting" which are scrolled
across.

While doing this I figured out a long standing problem with the panel
"light" MQTT topics: it never came on at startup because the OFF command
topic was retained. Insetad the light now honours brightness and on/off
topics, which may be retained. HA is now told to send the on/off after
setting the brightness and both topics may be retained.

Switch to ESP-IDF 5.5.1, which seems to have given me a new sdkconfig.
Actually on the subject of ESP32, I have an occasional problem with the
MQTT connection not establishing on startup. Might have to introduce a
horrible workaround for that. :(