Many, many small cleanups
Mostly just reordering #includes and fixing compiler warnings, and
renaming all classes and types to be CamelCase as after working on
the sensor classes I think it looks neater.

Also made it is the timezone can be passed in via TZ in local.cmake,
and the buzzer.cpp is always built but wont do anything useful on
ESP32 at the moment. To use it on a Pico board, set the BUZZER_PIN
in the board definition file.