Fix a stack overflow on ESP32
Even though a buzzer isn't yet useable on an ESP32 board, the buzzer
task still runs as a target for MQTT posts. ESP32 seems to be more
stack-demanding then Pico and I had overflows in the buzzer task when
doing notifications. Move the RTTTL object into the heap to fix them.